RDA-732 support DRAFTS to import and index relationships

+++ mycelium/src/main/java/ardc/cerium/mycelium/rifcs/executor/PrimaryKeyDeletionExecutor.java
@@ -63,6 +63,11 @@ public class PrimaryKeyDeletionExecutor extends Executor {
 
 		// remove all PrimaryKey edge to the key
 		String key = sideEffect.getKey();
+		Vertex roVertex = getMyceliumService().getRegistryObjectVertexFromKey(key);
+		if(roVertex == null){
+			log.debug("Vertex with PrimaryKey:{} Doesn't exist",  key);
+			return;
+		}
 		getMyceliumService().getGraphService().deletePrimaryKeyEdge(key);
 
 		// remove all PrimaryKey edge from/to the roVertex
